refactor: additional cleanup for derivation
