+++ frontend/src/metabase/lib/colors/constants/lightness-stops.ts
@@ -1,61 +1,5 @@
 import type { MetabaseColorKey } from "../types";
-import type { LightnessStop } from "../types/lightness-stops";
-
-/**
- * A step derivation can be:
- * - Assign a fixed step number (5, 10, 20, ... 110)
- * - Use a relative offset from the detected step (e.g. offset: -2)
- * - A special accessor function name
- */
-export type StepDerivation =
-  | LightnessStop
-  | { offset: number }
-  | { accessor: "accessibleText" }
-  | { accessor: "accessibleText"; offsetFromResult: number }
-  | { accessor: "accessibleBackground" };
-
-/**
- * Alpha derivation uses alpha or alphaInverse stops
- */
-export type AlphaDerivation =
-  | { alpha: LightnessStop }
-  | { alphaInverse: LightnessStop };
-
-export type Derivation = StepDerivation | AlphaDerivation;
-
-/**
- * Conditional derivation based on theme and brand lightness
- */
-export interface ConditionalDerivation {
-  darkTheme?: {
-    darkBrand?: Derivation; // brand step >= 50
-    lightBrand?: Derivation; // brand step < 50
-    default?: Derivation; // fallback if no brand condition
-  };
-  lightTheme?: {
-    darkBrand?: Derivation;
-    lightBrand?: Derivation;
-    default?: Derivation;
-  };
-  // Simple case: same for both themes
-  default?: Derivation;
-}
-
-export type DerivationRule = Derivation | ConditionalDerivation;
-
-// =============================================================================
-// Constants
-// =============================================================================
-
-// Default colors from light theme (used for dark mode detection when colors aren't provided)
-// eslint-disable-next-line no-color-literals
-export const DEFAULT_LIGHT_BACKGROUND = "#ffffff";
-// eslint-disable-next-line no-color-literals
-export const DEFAULT_LIGHT_TEXT = "#303030";
-
-// =============================================================================
-// Derivation Maps
-// =============================================================================
+import type { Derivation, DerivationRule } from "../types/lightness-stops";
 
 /**
  * Background color derivations from background-primary
